+++ lib/osf-components/app/components/scheduled-banner/x-maybe-link/component.ts
@@ -1,0 +1,1 @@
+export { default } from 'osf-components/components/scheduled-banner/x-maybe-link/component';