+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Models:
+    - `banner` - used to fetch `/_/banners/current/` from the API
+- Components:
+    - `scheduled-banner` - display the "current" banner on the landing page(s)
+
 ## [0.5.0] - 2018-06-29
 ### Added
 - Routes: